newsapi java : reorganize packages
